requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
celery==5.4.0
redis==5.2.0
flower==2.0.1
pydantic-settings==2.6.1

# Image Processing
Pillow==10.4.0
img2pdf==0.5.1
pdf2image==1.17.0

# PDF and Document Processing
pdf2docx==0.5.8
python-docx==1.1.2
pypdf2==3.0.1

# File Type Detection
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# AWS S3 (Optional)
boto3==1.35.0
botocore==1.35.0

# Utilities
python-dotenv==1.0.1
